graceful job cancelation for paralleljobhandler and run in container

+++ hither/paralleljobhandler.py
@@ -1,3 +1,6 @@
+import os
+import signal
+import tempfile
 from typing import List, Dict, Any
 import time
 import multiprocessing
@@ -22,7 +25,13 @@ class ParallelJobHandler(BaseJobHandler):
         import kachery as ka
         pipe_to_parent, pipe_to_child = multiprocessing.Pipe()
         serialized_job = job._serialize(generate_code=(job._container is not None))
-        process = multiprocessing.Process(target=_pjh_run_job, args=(pipe_to_parent, serialized_job, ka.get_config()))
+
+        # Note that cancel_filepath will only have an effect if we are running this in a container
+        if job._container is not None:
+            cancel_filepath = f'{tempfile.gettempdir()}/pjh_cancel_job_{job._job_id}.txt'
+        else:
+            cancel_filepath = None
+        process = multiprocessing.Process(target=_pjh_run_job, args=(pipe_to_parent, cancel_filepath, serialized_job, ka.get_config()))
         self._processes.append(dict(
             job=job,
             process=process,
@@ -34,15 +43,35 @@ class ParallelJobHandler(BaseJobHandler):
         for p in self._processes:
             if p['job']._job_id == job_id:
                 if p['pjh_status'] == JobStatus.RUNNING:
-                    pp = p['process']
-                    print(f'ParallelJobHandler: Terminating process.')
-                    pp.terminate()
-                    pp.join()
-                p['job']._result = None
-                p['job']._status = JobStatus.ERROR
-                p['job']._exception = JobCancelledException('Job canceled')
-                p['job']._runtime_info = None
-                p['pjh_status'] = JobStatus.ERROR
+                    # Note that cancel_filepath will only have an effect if we are running this in a container
+                    if p['job']._container is not None:
+                        print(f'ParallelJobHandler: Stopping process by writing cancel file.')
+                        cancel_filepath = f'{tempfile.gettempdir()}/pjh_cancel_job_{job_id}.txt'
+                        with open(cancel_filepath + '.tmp', 'w') as f:
+                            f.write('cancel.')
+                        os.rename(cancel_filepath+'.tmp', cancel_filepath)
+                    else:
+                        print(f'ParallelJobHandler: Terminating job.')
+                        pp = p['process']
+                        pp.terminate()
+                        pp.join()
+                        p['job']._result = None
+                        p['job']._status = JobStatus.ERROR
+                        p['job']._exception = JobCancelledException('Job cancelled')
+                        p['job']._runtime_info = None
+                        p['pjh_status'] = JobStatus.ERROR
+                    # if pp.is_alive():
+                    #     print('--- x2')
+                    #     pp.join(timeout=2)
+                    #     print('--- x3')
+                    # print(f'ParallelJobHandler: Process stopped.')
+                else:
+                    # TODO: Consider if existing ERROR or FINISHED status should change this behavior 
+                    p['job']._result = None
+                    p['job']._status = JobStatus.ERROR
+                    p['job']._exception = JobCancelledException('Job cancelled')
+                    p['job']._runtime_info = None
+                    p['pjh_status'] = JobStatus.ERROR
     
     def iterate(self):
         if self._halted:
@@ -74,11 +103,12 @@ class ParallelJobHandler(BaseJobHandler):
         
         time.sleep(0.02)
 
-def _pjh_run_job(pipe_to_parent: Connection, serialized_job: Any, kachery_config: dict) -> None:
+def _pjh_run_job(pipe_to_parent: Connection, cancel_filepath: str, serialized_job: Any, kachery_config: dict) -> None:
     import kachery as ka
     ka.set_config(**kachery_config)
     job = hi._deserialize_job(serialized_job)
-    job._execute()
+    # Note that cancel_filepath will only have an effect if we are running this in a container
+    job._execute(cancel_filepath=cancel_filepath)
     ret = dict(
         result=job._result,
         status=job._status,
